fix(sendit): un COD avec des centimes cassait toute la synchro
« invalid input syntax for type integer: "463.5" » — le colis DH20FF445D7
(Imane khattabi, 463,50 MAD) faisait echouer le bouton « Sync statut/COD »
a chaque clic, et tous les colis traites APRES lui n'etaient jamais
synchronises.

La cause n'est pas le montant mais l'inference de type de PostgreSQL. Dans
`CASE WHEN $5 > 0 THEN $5 ELSE total END`, le litteral `0` est un entier :
Postgres en deduit que $5 est un entier, et refuse 463.5 — alors que la
colonne cible est numeric(10,2) et l'aurait parfaitement accepte. Le
message d'erreur designe donc le montant alors que le probleme est la
requete.

Corrige par un cast explicite `$5::numeric` sur chaque usage (et $4 pour
les frais, par coherence entre les deux branches prepaye / COD).

Reproduit puis verifie en base :
- `SELECT CASE WHEN $1 > 0 ...` avec 463.5 -> meme erreur, a l'identique
- la meme avec `$1::numeric` -> passe
- la requete complete de sync, jouee sur une commande de test en
  transaction annulee -> total 463.5, codAmount 463.50, paidAmount 463.50,
  paymentStatus PAID

Etat verifie apres correction : 0 commande livree chez Sendit et non
livree dans le BOS. Restent 4 colis ou Sendit annonce un montant de 0 —
le garde-fou `> 0` protege alors les valeurs du BOS, c'est voulu.

+++ app/api/ops/sendit/staging/route.ts
@@ -149,7 +149,7 @@ export async function POST(req: NextRequest) {
                  "senditStatus" = $2::text,
                  "deliveryStatus" = $2::varchar,
                  status = $3::"OrderStatus",
-                 "actualDeliveryCost" = $4,
+                 "actualDeliveryCost" = $4::numeric,
                  "codAmount" = NULL,
                  "deliveredAt" = CASE
                    WHEN $3 = 'DELIVERED' AND NULLIF($5::text, '') IS NOT NULL
@@ -166,17 +166,22 @@ export async function POST(req: NextRequest) {
                  "senditStatus" = $2::text,
                  "deliveryStatus" = $2::varchar,
                  status = $3::"OrderStatus",
-                 "actualDeliveryCost" = $4,
-                 total = CASE WHEN $5 > 0 THEN $5 ELSE total END,
-                 "codAmount" = CASE WHEN $5 > 0 THEN $5 ELSE "codAmount" END,
-                 "paidAmount" = CASE WHEN $3 = 'DELIVERED' AND $5 > 0 THEN $5 ELSE "paidAmount" END,
+                 "actualDeliveryCost" = $4::numeric,
+                 -- $5::numeric est OBLIGATOIRE. Sans cast, « $5 > 0 » fait deduire a
+                 -- Postgres que le parametre est un ENTIER (a cause du litteral 0) :
+                 -- tout COD avec des centimes echouait alors sur
+                 -- « invalid input syntax for type integer: "463.5" », et la synchro
+                 -- entiere s'arretait.
+                 total = CASE WHEN $5::numeric > 0 THEN $5::numeric ELSE total END,
+                 "codAmount" = CASE WHEN $5::numeric > 0 THEN $5::numeric ELSE "codAmount" END,
+                 "paidAmount" = CASE WHEN $3 = 'DELIVERED' AND $5::numeric > 0 THEN $5::numeric ELSE "paidAmount" END,
                  "paidAt" = CASE
                    WHEN $3 = 'DELIVERED' AND NULLIF($6::text, '') IS NOT NULL
                      THEN ($6::timestamp AT TIME ZONE 'Africa/Casablanca')
                    ELSE "paidAt"
                  END,
                  "paymentReference" = CASE WHEN $3 = 'DELIVERED' THEN COALESCE("paymentReference", $1) ELSE "paymentReference" END,
-                 "paymentStatus" = CASE WHEN $3 = 'DELIVERED' AND $5 > 0 THEN 'PAID' ELSE "paymentStatus" END,
+                 "paymentStatus" = CASE WHEN $3 = 'DELIVERED' AND $5::numeric > 0 THEN 'PAID' ELSE "paymentStatus" END,
                  "deliveredAt" = CASE
                    WHEN $3 = 'DELIVERED' AND NULLIF($6::text, '') IS NOT NULL
                      THEN ($6::timestamp AT TIME ZONE 'Africa/Casablanca')
